# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_Person.age-or-name_": {
				"properties": {
					"age": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"age",
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"PersonCreationParams": {
				"$ref": "#/components/schemas/Pick_Person.age-or-name_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "hngx-stage-one",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PersonCreationParams"
							}
						}
					}
				}
			}
		},
		"/api/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"age": {
													"type": "number",
													"format": "double"
												},
												"name": {
													"type": "string"
												},
												"user_id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"age",
												"name",
												"user_id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"properties": {
												"age": {
													"type": "number",
													"format": "double"
												},
												"name": {
													"type": "string"
												},
												"user_id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"age",
												"name",
												"user_id"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PersonCreationParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"user": {
													"properties": {
														"age": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														},
														"user_id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"age",
														"name",
														"user_id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"user",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"user": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}